ASTERIX: Widen a loop variable
Fix

    /wireshark/epan/dissectors/packet-asterix.c:61495:35: error: loop variable has narrower type 'uint8_t' than iteration's upper bound 'const unsigned int' [bugprone-too-small-loop-variable,-warnings-as-errors]
     61495 |             for (i = 0, size = 0; i < field->header_length; i++) {
           |                                   ^

(cherry picked from commit 6a6f5da68ef77ea285acaf6d75142b6c9aa5d62f)

Conflicts:
	epan/dissectors/packet-asterix.c
	tools/asterix/packet-asterix-template.c

+++ epan/dissectors/packet-asterix.c
@@ -52728,7 +52728,7 @@ static int asterix_field_length (tvbuff_t *tvb, unsigned offset, const AsterixFi
 {
     unsigned size;
     uint64_t count;
-    uint8_t i;
+    unsigned i;
 
     size = 0;
     switch(field->type) {
